requirements.txt
PyGithub
google-generativeai
python-dotenv
pydantic
requests
tqdm
